Extract authenticate into AuthService
